specify types on test

+++ tests/python/relay/test_op_grad_level3.py
@@ -121,8 +121,8 @@ def test_gather_nd_grad():
     data = relay.var("data", relay.TensorType((2, 3), "float64"))
     indices = relay.var("indices", relay.TensorType((2, 4), "int64"))
     fwd = relay.Function([data, indices], relay.gather_nd(data, indices))
-    data_np = np.random.rand(2, 3)
-    indices_np = np.array([[0, 1, 1, 0], [0, 1, 2, 1]])
+    data_np = np.random.rand(2, 3).astype("float64")
+    indices_np = np.array([[0, 1, 1, 0], [0, 1, 2, 1]]).astype("int64")
     check_grad(fwd, inputs=[data_np, indices_np], test_inputs=indices_np)
